# openapi.yaml
openapi: 3.0.1
info:
  title: Queuing
  description: |-
    Queuing API 사용 방법과 예시를 담은 문서예요.

    API 사용 중 오류가 발생하거나, 질문이 있다면 새벽이라도 부담스러워 하지 마시고 연락주세요!
  version: 1.0.0
...
paths:
...
  /api/v1/friends:
    get:
      summary: 친구 목록 조회
      deprecated: false
      description: ''
      tags:
        - 친구
      parameters:
        - name: lastId
          in: query
          description: 마지막으로 조회한 친구 ID (Pagination)
          required: false
          schema:
            type: integer
            format: int64
            example: 10
        - name: size
          in: query
          description: 조회할 개수
          required: false
          schema:
            type: integer
            default: 20
            example: 20
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListFriendResponseBody'
              example:
                result:
                  items:
                    - id: 123
                      nickname: 친구1
                      slug: friend-1-slug
                      profileImageUrl: https://example.com/1.jpg
                    - id: 124
                      nickname: 친구2
                      slug: friend-2-slug
                      profileImageUrl: null
                  hasNext: true
          headers: {}
          x-apidog-ordering: 0
      security: []
      x-apidog-folder: 루트/친구
      x-apidog-status: released
      x-run-in-apidog: https://app.apidog.com/web/project/1164289/apis/api-26074171-run
  /api/v1/friends/{targetSlug}:
    delete:
      summary: 친구 삭제
      deprecated: false
      description: ''
      tags:
        - 친구
      parameters:
        - name: targetSlug
          in: path
          description: 삭제할 친구의 슬러그
          required: true
          example: ''
          schema:
            type: string
            example: target-user-slug
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponseBody'
              example:
                result: true
          headers: {}
          x-apidog-ordering: 0
        '404':
          description: 사용자 또는 친구 관계를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                UserNotFound:
                  summary: UserNotFound
                  value:
                    error:
                      statusCode: 404
                      code: user.not-found
                      message: 사용자 정보가 존재하지 않아요.
                FriendRequestNotFound:
                  summary: FriendRequestNotFound
                  value:
                    error:
                      statusCode: 404
                      code: friend.request-not-found
                      message: 친구 요청을 찾을 수 없어요.
          headers: {}
          x-apidog-ordering: 1
      security: []
      x-apidog-folder: 루트/친구
      x-apidog-status: released
      x-run-in-apidog: https://app.apidog.com/web/project/1164289/apis/api-26074172-run
  /api/v1/friend-requests:
    post:
      summary: 친구 요청
      deprecated: false
      description: ''
      tags:
        - 친구
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendFriendRequest'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponseBody'
              example:
                result: true
          headers: {}
          x-apidog-ordering: 0
        '400':
          description: 잘못된 요청 (자신에게 요청 등)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                InvalidRequest:
                  summary: InvalidRequest
                  value:
                    error:
                      statusCode: 400
                      code: friend.invalid-request
                      message: 허용되지 않은 친구 요청이에요.
          headers: {}
          x-apidog-ordering: 1
        '404':
          description: 대상을 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                UserNotFound:
                  summary: UserNotFound
                  value:
                    error:
                      statusCode: 404
                      code: user.not-found
                      message: 사용자 정보가 존재하지 않아요.
          headers: {}
          x-apidog-ordering: 2
        '409':
          description: 이미 친구 요청을 보냈거나 친구인 경우
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                AlreadyExists:
                  summary: AlreadyExists
                  value:
                    error:
                      statusCode: 409
                      code: friend.already-exists
                      message: 이미 친구 관계거나 요청 중이에요.
          headers: {}
          x-apidog-ordering: 3
      security: []
      x-apidog-folder: 루트/친구
      x-apidog-status: released
      x-run-in-apidog: https://app.apidog.com/web/project/1164289/apis/api-26074173-run
  /api/v1/friend-requests/{requestId}:
    patch:
      summary: 친구 요청 상태 변경 (수락 등)
      deprecated: false
      description: ''
      tags:
        - 친구
      parameters:
        - name: requestId
          in: path
          description: 요청 ID
          required: true
          example: 0
          schema:
            type: integer
            format: int64
            example: 1
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/FriendRequestUpdate'
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BooleanResponseBody'
              example:
                result: true
          headers: {}
          x-apidog-ordering: 0
        '404':
          description: 요청을 찾을 수 없거나 권한이 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                RequestNotFound:
                  summary: RequestNotFound
                  value:
                    error:
                      statusCode: 404
                      code: friend.request-not-found
                      message: 친구 요청을 찾을 수 없어요.
          headers: {}
          x-apidog-ordering: 1
      security: []
      x-apidog-folder: 루트/친구
      x-apidog-status: released
      x-run-in-apidog: https://app.apidog.com/web/project/1164289/apis/api-26074174-run
...
  /api/v1/rooms:
    post:
      summary: 방 생성
      deprecated: false
      description: ''
      tags:
        - 방
      parameters: []
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateRoomRequest'
      responses:
        '201':
          description: 생성됨 (Location 헤더에 생성된 방의 슬러그가 포함됨)
          headers:
            Location:
              required: false
              description: '생성된 방의 슬러그 (예: /api/v1/rooms/my-awesome-room)'
              schema:
                type: string
          x-apidog-ordering: 0
        '400':
          description: 입력 값이 잘못됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                InvalidInput:
                  summary: InvalidInput
                  value:
                    error:
                      statusCode: 400
                      code: common.invalid-input
                      message: 입력 값이 잘못되었어요.
          headers: {}
          x-apidog-ordering: 1
        '404':
          description: 사용자 정보를 찾을 수 없음
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                UserNotFound:
                  summary: UserNotFound
                  value:
                    error:
                      statusCode: 404
                      code: user.not-found
                      message: 사용자 정보가 존재하지 않아요.
          headers: {}
          x-apidog-ordering: 2
      security: []
      x-apidog-folder: 루트/방
      x-apidog-status: released
      x-run-in-apidog: https://app.apidog.com/web/project/1164289/apis/api-26074176-run
    get:
      summary: 방 목록 조회
      deprecated: false
      description: 최근 생성된 순으로 방 목록을 조회해요.
      tags:
        - 방
      parameters:
        - name: size
          in: query
          description: |-
            조회할 방 개수를 입력해주세요.
            입력하지 않아도 기본값으로 30개를 조회해요.
          required: false
          schema:
            type: integer
            default: 30
            minimum: 1
            maximum: 100
            example: 30
        - name: lastId
          in: query
          description: |-
            마지막으로 조회한 방 ID를 입력해주세요.
            입력한 ID보다 이전에 만들어진 방이 조회돼요.
          required: false
          schema:
            type: integer
            format: int64
            minimum: 1
            example: 100
      responses:
        '200':
          description: 성공
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ListRoomResponseBody'
              example:
                result:
                  rooms:
                    - id: 100
                      slug: study-room-1
                      title: 같이 공부해요
                      isPrivate: false
                      createdAt: '2023-10-27T10:00:00Z'
                      tags:
                        - slug: lo-fi
                          name: Lo-fi
                    - id: 99
                      slug: kpop-party
                      title: K-POP 파티
                      isPrivate: true
                      createdAt: '2023-10-27T09:30:00Z'
                      tags:
                        - slug: k-pop
                          name: K-POP
                  hasNext: true
          headers: {}
          x-apidog-ordering: 0
        '400':
          description: 입력 값이 잘못됨
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FailedResponseBody'
              examples:
                InvalidInput:
                  summary: InvalidInput
                  value:
                    error:
                      statusCode: 400
                      code: common.invalid-input
                      message: 입력 값이 잘못되었어요.
          headers: {}
          x-apidog-ordering: 1
      security: []
      x-apidog-folder: 루트/방
      x-apidog-status: released
      x-run-in-apidog: https://app.apidog.com/web/project/1164289/apis/api-26074177-run
components:
  schemas:
    BooleanResponseBody:
      type: object
      properties:
        result:
          type: boolean
          description: 성공 여부
          example: true
      x-apidog-orders:
        - result
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    FailedResponseBody:
      type: object
      properties:
        error:
          type: object
          properties:
            statusCode:
              type: integer
              description: HTTP 상태 코드
              example: 400
            code:
              type: string
              description: 에러 코드
              example: common.invalid-input
            message:
              type: string
              description: 에러 메시지
              example: 입력 값이 잘못되었어요.
            fieldErrors:
              type: array
              description: 필드별 검증 오류 목록 (Validation 실패 시)
              items:
                type: object
                properties:
                  field:
                    type: string
                    description: 오류가 발생한 필드명
                  value:
                    type: string
                    description: 거부된 값
                  reason:
                    type: string
                    description: 오류 원인
                x-apidog-orders:
                  - field
                  - value
                  - reason
                x-apidog-ignore-properties: []
          x-apidog-orders:
            - statusCode
            - code
            - message
            - fieldErrors
          x-apidog-ignore-properties: []
      x-apidog-orders:
        - error
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
...
    ListFriendResponseBody:
      type: object
      properties:
        result:
          $ref: '#/components/schemas/ListFriendResponse'
      x-apidog-orders:
        - result
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
...
    ListRoomResponseBody:
      type: object
      properties:
        result:
          $ref: '#/components/schemas/ListRoomResponse'
      x-apidog-orders:
        - result
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
...
    SendFriendRequest:
      type: object
      required:
        - targetSlug
      properties:
        targetSlug:
          type: string
          description: 친구 요청을 보낼 상대방의 Slug
          example: friend-user-slug
      x-apidog-orders:
        - targetSlug
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    FriendRequestUpdate:
      type: object
      required:
        - status
      properties:
        status:
          $ref: '#/components/schemas/FriendStatus'
      x-apidog-orders:
        - status
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    FriendStatus:
      type: string
      description: 친구 요청 상태
      enum:
        - PENDING
        - ACCEPTED
      example: ACCEPTED
      x-apidog-folder: ''
    FriendResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 친구 관계 ID
          example: 123
        nickname:
          type: string
          description: 친구 닉네임
          example: 친구닉네임
        slug:
          type: string
          description: 친구 Slug
          example: friend-user-slug
        profileImageUrl:
          type: string
          description: 친구 프로필 이미지 URL
          nullable: true
          example: https://example.com/friend.jpg
      x-apidog-orders:
        - id
        - nickname
        - slug
        - profileImageUrl
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    ListFriendResponse:
      type: object
      properties:
        items:
          type: array
          description: 친구 목록
          items:
            $ref: '#/components/schemas/FriendResponse'
        hasNext:
          type: boolean
          description: 다음 페이지 존재 여부
          example: true
      x-apidog-orders:
        - items
        - hasNext
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    CreateRoomRequest:
      type: object
      required:
        - title
      properties:
        title:
          type: string
          description: 방 제목
          example: 같이 공부해요
        password:
          type: string
          description: 방 비밀번호 (선택)
          nullable: true
          example: '1234'
        tags:
          type: array
          description: 태그 목록 (Slug)
          items:
            type: string
            example: lo-fi
      x-apidog-orders:
        - title
        - password
        - tags
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    RoomSummaryResponse:
      type: object
      properties:
        id:
          type: integer
          format: int64
          description: 방 ID
          example: 1
        slug:
          type: string
          description: 방 Slug
          example: study-room-1
        title:
          type: string
          description: 방 제목
          example: 같이 공부해요
        isPrivate:
          type: boolean
          description: 비공개 방 여부
          example: false
        createdAt:
          type: string
          format: date-time
          description: 방 생성 일시
          example: '2023-10-27T10:00:00Z'
        tags:
          type: array
          description: 방 태그 목록
          items:
            $ref: '#/components/schemas/MusicTagResponse'
      x-apidog-orders:
        - id
        - slug
        - title
        - isPrivate
        - createdAt
        - tags
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    ListRoomResponse:
      type: object
      properties:
        rooms:
          type: array
          description: 방 목록
          items:
            $ref: '#/components/schemas/RoomSummaryResponse'
        hasNext:
          type: boolean
          description: 다음 페이지 존재 여부
          example: true
      x-apidog-orders:
        - rooms
        - hasNext
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
    MusicTagResponse:
      type: object
      properties:
        slug:
          type: string
          description: 태그 Slug
          example: k-pop
        name:
          type: string
          description: 태그 이름
          example: K-POP
      x-apidog-orders:
        - slug
        - name
      x-apidog-ignore-properties: []
      x-apidog-folder: ''
...
  securitySchemes: {}
servers:
  - url: https://api.queuing.patulus.com
    description: 개발
security: []